govim: refactor to have an explicit event queue
Some calls from govim to Vim will result in events. If such a call
happens in a function/command callback, then we will deadlock because
the events will not be able to "fire".

Instead we split out a separate event queue. This allows calls made from
govim to Vim to result in events and for that event queue to be flushed
before we return to the caller.