README: improve readme and sign-posting (#79)
The biggest challenge with READMEs is that we have three things to
describe/explain:

1. the Vim8 plugin hosted at https://github.com/myitcv/govim
2. the github.com/myitcv/govim package, for people interested in writing
Vim8 plugins in Go
3. command github.com/myitcv/govim/cmd/govim, which is the main part of
the plugin described in point 1

It is expected/hoped that the vast majority of people visiting this
repository will be solely interested in point 1. Hence the README and
sign-posting needs to be heavily geared towards them.