govim: allow autocmds to declare expressions
Allow autocmds to declare expressions that will be evaluated when the
autocmd fires. This is important for autocmds like BufCreate where the
current buffer file name does not necessarily match the file name of the
buffer being manipulated. Hence the special expansion <afile> (et al)
are defined.

This also saves us a round-trip when updating buffer contents.

Closes #57