cmd/govim: ensure that we correctly add a newline to end of file
Our method of getting the entire buffer from Vim was failing to add a
trailing (end of file) newline (\n). Hence we were passing the file
contents minus the trailing \n to gopls. Which results in a bug in calls
to Formatting/CodeAction, but is almost certainly also invalid.

You have to go out of your way to avoid writing a trailing \n in Vim,
and for .go files it's almost certainly the case that the user wants to
write one.

We can revisit this more precisely if there is a pressing case for _not_
having a trailing \n.