govim: add Version() and Flavor() to Govim interface
This allows a plugin author (e.g. github.com/myitcv/govim/cmd/govim) to
handle differences in flavours and versions of Vim.

We also put in place the beginnings of a code-generated build matrix so
that Go code, Travis config and shell scripts are consistent as far as
default Go and Vim versions are concerned (makes ad hoc testing easier).
This is a simple starting point and very much a WIP. It's anticipated
that at a later date we will switch all scripts to be simply Go code.

Travis now also uses _scripts/buildGovimImage.sh for consistency with
the developer.

VIM_FLAVOR is now a fully required environment variable for non-Docker
based testing. This ensures that when a developer is running tests
locally they are forced to specify whether they want to test vim/gvim
etc locally.

_scripts/buildGovimImage.sh and _scripts/runDockerRun.sh have been
improved to make it easier for ad hoc testing by folks not interested in
installing vim/gvim/nvim (e.g. the gopls team). Just like on CI, such a
developer can run:

  $ _scripts/buildGovimImage.sh # ensure local govim image is current
  $ _scripts/runDockerRun.sh

(note VIM_FLAVOR does not need to be set; it defaults to vim) to perform
a CI-esque test run on the working tree. The first step simply ensures
your local govim Docker image is current; it need not be repeated before
every call to _scripts/runDockerRun.sh.

In all _scripts scripts, VIM_VERSION is used to refer to the version of
VIM_FLAVOR, i.e. it is either the version of vim, gvim, or nvim (when we
add that support). Confusingly, everywhere else, e.g. MAX_VIM_VERSION or
VIM_VERSIONS, the "VIM" part of the name actually refers to Vim; not
Gvim or Neovim.