cmd/govim: work around complete_watched flake (#230)
It's not clear right now whether gopls will/won't block in this test
case, i.e. after the didOpen is received by gopls, will a subsequent
completion request block until the processing of the new file is done.

Hence we pessimistically wait until we receive the diagnostics for the
file.