cmd/govim: fix completion
https://go-review.googlesource.com/c/tools/+/176958 changed the
TextEdits returned from gopls to not be the trimmed version of the
label. This meant that we were effectively appending the entire
completion text after the prefix.

This wasn't caught in the completion tests because there we were
attempting a "naked" completion with no prefix.

Fix by fixing the test (which confirmed the bug) then fixing the results
returned from gopls.

Seemingly, a list of completion items from gopls can specify a start
position per item. This does not appear to be supported by Vim (see
:help complete-functions)