govim: initial cut of proper shutdown sequence
Because Vim and gopls are both sources of events as far as govim is
concerned, we effectively have a race when it comes to shutdown. Vim
will always be the one initiating the shutdown, but that notice can
happen at exactly the same time as receiving a message from gopls that
requires us to do something in Vim.

Hence shutdown is initiated by Vim by closing the channel. This allows
the read loop in govim to exit cleanly.

But at this point writes to the connection (off the back of events from
gopls for example) will fail.

Therefore, we define that any error whilst attempting to write to the
connection is indicative of the fact that we (govim) are shutting down.

When we detect we are shutting down, we panic with a known error. This
allows us to cleanly and quickly unwind in a well defined manner.