govim: remove viewport subscription code
Vim is not currently in a position to be able to accurately push
viewport changes to govim. Hence, we might as well drop this code, not
least because it clutters the log messages having the viewport changes
being pushed the whole time. But also because we don't have anything in
govim that relies on knowing when the viewport changes (that is #24).

This should also make adding the Neovim support easier.