cmd/govim: better model for handling syntax errors
Syntax errors in a file mean that a call to any of the gopls methods
that format the code will fail. This is confirmed by those methods
returning errors.

Currently we return that error to Vim, meaning the error is thrown. Not
only that, but, if the format function is called as part of
format-on-save, then the save is aborted.

Per #274, this is not ideal.

Instead, we choose Option 2 in #274 for now:

* allow the save
* do not echo/throw any error
* rely on the user spotting the syntax error in the quickfix window

We might follow this up with a further PR that automatically shows the
quickfix window if there are any errors in it, but not for now.

Fixes #274