cmd/govim: de-dupe PublishDiagnostics notifications from gopls
This helps to throttle the number of changes we try to push through to
the quickfix window.

However, https://github.com/golang/go/issues/32443 remains an issue in
as much as incomplete diagnostics are sent to the client. i.e. we can't
be sure to receive the most recent diagnostics for the file which we are
editing (and in which we know, for example, we are making errors).