govim: use v:longversion in Vim for more precise version number
Our parsing of the output of :version is hacky at best, broken at worst
as was recently demonstrated via the locale bug.

As of v8.1.1526, Vim now has v:longversion, from which we can precisely
get major, minor and patch.

When our minimum required version of Vim is >= v8.1.1526, we can drop
the old parsing logic