govim: improve testdriver debugging logging (#236)
This makes it easier to debug issues with the govim tests, by allowing
users to log more information with a general -debugLog flag. This
enables Vim debug logging, and sends Vim startup output to the normal
Govim log. If Vim fails to start, it prints out the contents of the log,
which should hopefully contain some useful information. It also prints a
full command to run Vim with the temporary home directory.

Since the call to NewTestDriver was getting a bit unwieldy with over 10
args, I refactored that out into a Config struct.