cmd/govim: use popup windows for hover details if Vim >= v8.1.1649
This PR is the first to take advantage of the new popup window support
in Vim/Gvim.

Balloons are quite limiting when it comes to display hover related
information:

* There are differences in the way in which text is rendered between
terminal Vim and Gvim
* We can highlight the contents of the ballooon
* ...

Therefore, it's considerably more flexible to use popup windows.

As part of this PR we define a condition on *vimstate that determines
whether we can use popup windows (for hover information).

We also add some test mechanics to dump the contents of popups visible
on screen. This allows us to test the command and mouse-triggered hover.

With HUGE thanks to @brammool for all of the changes that have lead us
to this point.

Fixes #283
Fixes #72