cmd/govim: support batching of function calls to Vim (#346)
This is a first cut of support for making batches of calls from govim to
Vim.

Batches are only supported in calls on *vimstate. A batch is started via
a *vimstate.BatchStart() call, and ends with a *vimstate.BatchEnd()
call. During a batch, only calls to *vimstate.ChannelCall() are allowed.
By definition, these calls always return nil. At the end of the batch,
the call to *vimstate.BatchEnd() returns an ordered list of results from
each of the calls made in the batch. An error during the evaluation of a
single call in the batch results in an error for the entire batch.

Fixes #345