cmd/govim: fix bug for large batches (#357)
Currently we define s:batchCall to take a variadic number of arguments.
This is unnecessary because it can take a single argument, a list. With
the signature being variadic, it also means we're likely to hit some
sort of Vim limit on the number of arguments that can be parsed.

This change changes the signature of that function to take a list,
thereby sidestepping all of these issues.