cmd/govim: provide an experiment config option for popup options
Whilst we try to establish the best defaults for popup windows in
various contexts (#360), let's provide experimental config options to
allow users to help with this experiment.

For example, to set the options for mouse-trigger hover popups, add the
following to your .vimrc:

call govim#config#Set("ExperimentalMouseTriggeredHoverPopupOptions", {
      \ "mousemoved": "any",
      \ "pos": "botleft",
      \ "line": -1,
      \ "col": 0,
      \ "moved": "any",
      \ "border": [],
      \ "wrap": v:false,
      \ "close": "click",
      \ "padding": [0, 1, 0, 1],
      \})

See the go doc for github.com/myitcv/govim/cmd/govim/config for more
details on what config options can be set.

See :help popup_create-arguments for details of the various options that
can be set for ExperimentalMouseTriggeredHoverPopupOptions and
ExperimentalCursorTriggeredHoverPopupOptions.

Whilst we are at it, also change the default popup to add some left and
right padding.