cmd/govim: improve efficiency of format changes
Currently we can end up triggering an listener_add callback for every
format change in a buffer. This is incredibly inefficient because we
will always be applying such changes in a batch.

Instead, drop the listener, apply the changes, read the entire buffer
contents, fire didChange, then re-add the listener.