ci: optimise CI runs for [WIP] PRs and add CI_MUST_RUN config option (#381)
When working on PRs, it is extremely expensive and time consuming to
have to wait for the entire build matrix to run.

This PR optimises this process by only running a build for matrix entry
where GO_VERSION and VIM_VERSION corresponds (for the value of
VIM_FLAVOR) to the value of MAX_GO_VERSION and MAX_${X}_VERSION
respectively.

This logic can be bypassed for certain PRs, that is to say the entire
build matrix will be run, by setting the CI_MUST_RUN_${PR_NUMBER}
environment variable. This will, most likely, be set in Travis rather
than a script in the project itself for ease of turning on/off this
switch.

As part of this optimisation spree, we also limit race tests to only
running on master.