cmd/govim: initial cut of GOVIMStringFn command
GOVIMStringFns is a command that replaces buffer text with the result of
applying a sequence of transformation functions. With no range, the
current line is used as input. That is to say the current line is used
as the argument to the sequence of transformation functions. Visual
selections can also be used to specify the range, however the command
has no meaning for visual blocks.

The command takes one or more arguments: the sequence of transformation
functions to apply. Each argument is a path to a transformation
function, e.g. strconv.Quote

Here is an example of applying multiple transformations to a visual
selection:

:'<,'>GOVIMStringFn crypto/sha256.Sum256 encoding/hex.EncodeToString

The goal with this command is to expose standard library functions to
help manipulate text. Wherever possible, functions will directly map to
their standard library equivalents, for example, strconv.Quote. In this
case, the format is $importpath.$function. In some situations, poetic
license may be required.

When invoking the command, tab completion can be used against the known
transformation functions.

For now we hand-define a couple of obvious transformation functions. In
future PRs we may look to code generate this mapping, as well as writing
string-based wrappers around things like sha256 hash calculations.

Closes #273