cmd/govim: move tools "vendor" to cmd/govim/internal/golang_org_x_tools
Sticking our 'vendor' of golang.org/x/tools directly into internal means
we cannot add internal packages of our own. In a later commit we move
the cmd/govim/types package to be such an internal package (because it
doesn't make sense for it to be anything else for now). So this is a
preparatory commit.