internal/plugin: tidy up shutdown errors
Currently we muddle up tomb.ErrDying errors with govim.ErrShuttingDown.
We only need one of these to indicate that govim is shutting down, and
it should be the latter; the former is an implementation detail.

As part of this tidy up, we also fix a bug in internal/plugin where we
were incorrectly wrapping as an ErrDriver an error caused by govim
shutting down. Given the simplification to now only use
govim.ErrShuttingDown to indicate the mode/event, we simply check we
aren't shutting down before otherwise creating a wrapped error.

Fixes #406