cmd/govim: expand the set of sensible out-of-the-box defaults
Expand the set of sensible out-of-the-box defaults we recommend in the
minimal .vimrc. In future, as previously discussed, we might want to
make these defaults a part of the govim plugin itself, but for now we
make them as recommendations in the minimal .vimrc

This change includes:

* turning on syntax highlighting for .go files
* turning on smart-autoindent
* setting sensible backspace behaviour

This should give a nicer experience for first-time Vim users.

Fixes #430