cmd/govim: fix setting GOBIN during install
Move the GOBIN and GO111MODULE environment variable sets into the system
call instead of setting them in VIMs shell. On my system setting GOBIN
in VIMs shell then calling system is still resulting in the shell run by
system having the GOBIN that's defined in my own environment. GOBIN is
defined in my shell rc file so it's possible when system is running the
command it's starting a new shell which is setting GOBIN again.

This install command should make less assumptions about the shell
underneath VIM and moving the environment variables into the system call
makes less since it overrides the variables for that single line being
executed rather than changing it for the VIM shell.