cmd/govim: de-dupe quickfix diagnostics and retain selected index (#476)
Currently, because of https://github.com/golang/go/issues/32443 we can
unnecessarily update the quickfix window with diagnostics. What's more,
when we update the quickfix list we lose the selected index; the first
row is selected when the quickfix list is reset.

This presents something of a problem when jumping to a quickfix error in
a file that is not currently open, where that error is not the first in
the quickfix list. Because jumping to that error means we trigger a
didOpen, which triggers all diagnostics to be sent again (even though
they haven't changed), which causes the first row to be selected again.
The user, understandably, get's very confused.

So for now, we use reflect.DeepEqual to de-dupe the diagnostics we
receive from gopls, pending a fix to
https://github.com/golang/go/issues/32443. We also keep track of the
last set error diagnostics, so that when setting new diagnostics we can
reselect the currently selected row. Or, if that row no longer exists,
select the first row. This changes the times that setqflist is called.
Hence, fix up tests which previously asserted based on this behaviour.

As part of this change we also retrigger clearing/population of the
quickfix list (and signs) as the relevant config changes.

Also slightly optimise the sign placement code to return early if signs
are disabled.