cmd/govim: do not create a popup if there is a zero Hover response (#493)
Currently, we show a popup in the case there is no error in the call to
Hover. However, we do not check that there is actually anthing to show.

Following https://go-review.googlesource.com/c/tools/+/186997/, the
Hover LSP method swallows errors where the user hovered over an area
where there is no information. So now we get back nil errors and empty
Hover{} values where before we got an error (which we simply logged).

Correct this by only showing a popup in case we have a non-zero value
Hover{}.