cmd/govim: provide advanced env var to use gopls from PATH
Per @cespare's motivation, this allows users to use a (company)
standardised version of gopls and upgrade a rate different to that of
govim. This is important for things like handling gofmt changes between
Go versions.

However, use of this environment variable comes with zero warranty,
because it's not guaranteed that govim's tests pass/work at all with
whatever version of gopls is found in PATH.

Fixes #440