cmd/govim: add info popup config to minimal.vimrc; add minimal.gvimrc
Now that Vim supports popups for the info of a completion candidate, and
that popup window can be positioned relative to the candidates menu,
provide a version-guarded suggestion in the minimal.vimrc.

Also create a minimal.gvimrc so that Gvim users don't miss out on this
suggestions. This will require a tweak to the wiki in a couple of
places.

Fixes #392