all: drop any references to now-old Vim versions
Following 63eec53 we now have a new minimum Vim version requirement for
govim: v8.1.1711. Therefore, any references to older versions are
defunct. Remove them, along with any conditions that were based on
previous versions.

As part of this we remove the now-legacy govim.ParseVimVersion to a test
file in testdriver because there doesn't appear to be a good way to get
the value of v:versionlong from the command line. Its accompanying test
moves with it.

We drop the condition test on the basis there's sufficient coverage by
virtue of conditions in testscript tests; no specific test is required.

Fixes #518