govim: use Vim's new SafeState* and state() to schedule work
As of e805441 we use a govim-local mechanism for scheduling work when it
is safe in Vim.

As of Vim v8.1.2056, we can leverage Vim-provided state/callbacks to
schedule this work, via the state() functions and SafeState*
autocommands.

In this change we leave the old govim-local mechanism in place for older
Vim versions, but switch to the state()/SafeState* approach for
sufficiently new versions.