cmd/govim: refactor diagnostics handling (#534)
To be able to add highlighting of diagnostics (over the entire range), diagnostics hovering, etc. we need to extract more things from the LSP diagnostics, like range and severity.

Much of the current logic take place inside the quickfix handling today, and the idea with this change is to move the generic stuff out from the quickfix specific code. That includes sign placement as well, but we still want to ensure a consistent state where signs always represents what is shown in the quickfix list.