cmd/govim: add suggested fixes via popup (#532)
This change adds "GOVIMSuggestedFixes" that will send a codeAction
request to gopls for all diagnostics on the current line. Each
diagnostic opens up a popup with suggested edits, visible only one
at a time. Cycle to the next diagnostic using <c-n>, and previous
using <c-p>.

Select suggested fix to apply using <enter> or close the menu
with <esc> to cancel.