cmd/govim: workaround flake in suggested_fixes test
For some reason the suggested_fixes test triggers some strange behaviour
in Vim. On some runs of the test the user never transitions to "busy" -
in these cases the test passes. On other runs the user does transition
to "busy" but then never transitions to idle, hence the test fails
(because we wait for the user to become idle before we update
diagnostics etc).

So for now we work around this by forcing the user to be idle at points
in the test when we know/expect they should be.