cmd/govim: keep track of if buffers are "Loaded" or not (#573)
To be able to use textprops for highlighting, we must be aware of
if a buffer is loaded or not. Text properties can't be added to
unloaded buffers, and trying to do so will throw an "invalid line"
error.

This change keeps track of if buffers are loaded or not.