cmd/govim: trigger gopls.DidConfigurationChange on config change
Now that https://go-review.googlesource.com/c/tools/+/206148 has landed
we can trigger a call to gopls to refetch configuration whenever the
user calls govim#config#Set() from Vim.

We will add proper tests in a future PR but from an intial test with
tweaking the staticcheck config on/off it doesn't look like gopls is
retriggering sending of diagnostics or actually turning on staticcheck
in the edge off -> on.