cmd/govim: add config for defining the gopls environment
gopls can be configured with an environment, an environment which is
passed to go/packages. For example, setting GOOS and GOARCH, or setting
GOFLAGS=-tags=other to set a build tag.

Therefore we define a govim config key, "GoplsEnv", for setting this
environment.

For example:

  govim#config#Set("GoplsEnv", {"GOFLAGS": "-modfile=go.local.mod"})

As part of this add a test that verifies the behaviour of the above
setting.

Setting GOFLAGS=-mod=readonly doesn't work; that is being tracked in
golang.org/issues/36789, and hence #555 needs to wait a bit longer. We
add a skipped test in any case.

Setting of build tags, i.e. GOFLAGS=-tags=other, does not currently
work; that will be followed up in a subsequent PR.