cmd/govim: refactor diagnostics handling to use lazy conversion (#585)
This change will use lazy conversion of incoming LSP diagnostics.

We use incoming diagnostics for both quickfix and sign placement,
and will also use them for highlighting. Vim handles each of them
differently when it comes to removing entities so we need a way
to get the latest diagnostics that doesn't perform conversion each
call.