cmd/govim: add goproxytest proxy server
Using github.com/rogpeppe/go-internal/goproxytest allows us to require
non standard library modules that are served from txtar archives
via a localhost proxy.

This is important for subsequent PRs where we need to verify certain
changes to go.mod files, and non standard library requirements are the
thing under test.

Making this change requires us to make a slight change to the
initialisation of the govim plugin so that we can pass the env from the
testdriver so that gopls is started with the correct GOPROXY.