cmd/govim: block govim initialisation until post Configuration call
As a temporary, fragile workaround for

This is a rather fragile workaround for
https://github.com/golang/go/issues/35817, we block the initialisation
of govim until we have received the Configuration call from gopls.  It's
fragile because we are relying on gopls not handling any requests until
the response to Configuration is received and processed. In practice
this appears to currently be the case but there is no guarantee of this
going forward. Rather we hope that a fix for
https://github.com/golang/go/issues/35817 lands sooner rather than later
at whic point this workaround can go.

We also use a lock here because, despite it appearing that will only be a single
Configuration call and that if there were more they would be serial, we can't rely on
this.

This is fragile and best and built on broken assumptions at worst. But
it should suffice for now.