cmd/govim: fix various diagnostic, quickfix and sign issues
This is a breaking change. The config setting QuickfixSigns is now
entirely independent of QuickfixAutoDiagnostics, i.e. it is possible to
place signs based on diagnostics independent of those diagnostics being
used to populate the quickfix window.

We also change the way in which we set signs in an attempt to be more
efficient. We now clear all signs and reset signs in a single batch.
This also means we no longer de-dupe multiple signs on the same line
(which has implications for the test golden files).

We also drop the old environment variable override for disabling signs;
they are now fully mature.