all: hide home directory in script tests
As described in:

  https://github.com/govim/govim/issues/603#issuecomment-560125390

we currently have a situation whereby the home directory sits within the
root directory of a script test. Given that most scripts also put files
in this root directory, i.e. the script root is also the main module
root, we have a situation where files in the home directory are
contained within the main module.

We do not set GOPATH explicitly as part of our tests, hence GOPATH
defaults to being a subdirectory within the home directory, i.e. also
part of the main module.

As we are now using goproxytest, this means modules we access end up
being expanded to $HOME/go/pkg/mod.

The file watcher that is part of govim is somewhat broken (#604) and we
end up watching changes to nested modules in $HOME/go/pkg/mod.

Whilst we have raised #604 in order to fix the file watching problem, we
can also fix this problem in the short term by defining HOME as:

  $script_root/.home

At the same time, we fix a bug in the watcher whereby we were not
correctly skipping hidden directories.

Fixes #603