govim: use a single tomb
In 0b4b110d we failed to spot that, despite the call stacks suggesting
as much, the panic was occurring because we were calling
(*gopkg.in/tomb.v2).Go from outside a tomb. Indeed there were multiple
tombs in play:

* one created by govim
* one created by cmd/govim or the testdriver

This is actually a throwback to a very old TODO that wasn't captured in
code or an issue, to unify the use of tombs. i.e. only ever have a
single one (because that's all we need). However we didn't do that, and
now we're getting bitten by it :)

The ultimate fix is #609, but for now unify to a single tomb and remove
the redundant debugging code.

Fixing this actually uncovered a couple of other bugs. Ultimately we
need to rewrite the shutdown handling logic, but for now we also fix up
those bugs.

Also remove some rather verbose logging from the testdriver.

Also add some failsafe code in the testdriver that handles errors/panics
from govim when executing script commands. There was a rather obscure
panic we were seeing during race tests that turned out to be gopls
exiting because of a data race, something that was getting obscured by
our tests panic-ing because govim was shutting down. Instead we now dump
the relevant log in case of such a panic, something which immediately
uncovered the data race panic in gopls. This should happen rarely so the
code/formatting etc isn't pretty but that really doesn't matter much.