tests: establish known quickfix buffer number in testscripts tests
Before a test script starts, the files from the txtar archive are
extracted. Then govim (and gopls) are started. In some rare cases, gopls
can end up calculating diagnostics before the testscript has started.
Which means that it can be the case, because govim and Vim are already up
and running, that the first buffer actually gets used for quickfix as
opposed to, say, the first file we open.

Instead below we copen then cclose in vimrc to reserve and establish a
definite buffer number for the quickfix window. The quickfix buffer will
always be buffer number 2 (because vim creates a buffer when opening, a
buffer that is then used for the first file we open).

Update various tests to reflect this change.