cmd/govim: add concept of user config in test overrides (#606)
Currently in our script tests we can specify the default config values
that will be used in the test. This overrides the cmd/govim default that
the end user will experience. Whilst this might be necessary in some
limited test situations, it's generally not what we neeed/intend. Rather
any config we're looking to supply in a scenario is generally an initial
override for the cmd/govim default.

Hence, we now all a test script to specify an explicit default via
default_config.json in a scenario directory (which would be nil, i.e. no
file), and we also allow for user_config.json to be the user's initial
overrides that are applied atop the defaults (whatever they may be).

As part of this we write a simple code generator to automate the
creation of the method:

  (*cmd/govim/config.Config).Apply(*cmd/govim/config.Config)

to avoid human mistakes further down the line.