testdriver: expose max errlogmatch wait time (#621)
Currently we have a default errlogmatch wait time of 30s which can be
overridden by the environment variable GOVIM_ERRLOGMATCH_WAIT. But this
value is internal to the testdriver package.

In test cases in subsequent PRs we need to assert that we do not receive
a notification from gopls. This obviously means we cannot use
errlogmatch to wait for a notification. Instead we need to wait for an
indeterminate period of time and then assert we did not receive a
notification in that time. We might as well therefore wait for the
default length of time that errlogmatch waits when looking for a match.

Hence we expose DefaultErrLogMatchWait from testdriver and expose it as
an environment variable DEFAULT_ERRLOGMATCH_WAIT in all scripts so that
they can simply:

    sleep $DEFAULT_ERRLOGMATCH_WAIT

and then use errlogmatch with a count or similar.