tests: add a -gopls flag for scenario tests in cmd/govim
For validating changes to gopls, it would be helpful to be able to run
the govim integration tests against a specific copy of the gopls binary.
Currently, the scenario tests run against a freshly built pinned version
of gopls.

A -gopls flag is added to allow specifying a path to a pre-existing copy
of gopls, for use in these scenario tests.