cmd/govim: add config ExperimentalTempModfile
ExperimentalTempModfile corresponds to the gopls config setting
"tempModfile" which controls whether a temporary modfile is used in
place of the main module's original go.mod file. When enabled, any
user-initiated changes (to .go files) that would otherwise have resulted
in changes to the original go.mod file, e.g. adding an import for a
package whose module is not listed as a requirement, get raised as
diagnostic warnings with suggested fixes which update the go.mod file.

Add a test that locks in the current behaviour.