cmd/govim: create a test with expectations for diagnostic notifications
Now that gopls no longer sends duplicate notifications for a file (put
another way, we only get diagnostic notifications for a file when those
diagnostics change) we need to lock in our expectations in a test.

Reason being, we do lots of errlogmatch for diagnostic notifications so
our tests might appear flaky if our expectations are wrong.