cmd/govim: highlight diagnostics (#579)
Currently diagnostics as errors and warnings are only placed
in the quickfix buffer, with a sign in the file buffer gutter
next to the affected line.
This change also adds in-code text property highlighting to
the entire range affected by the diagnostic. It is combined
with existing syntax highlight.

Diagnostic highlighting can be disabled by setting the config
HighlightDiagnostics to 0 (i.e false).

Note that this is a breaking change if you have overridden
any of the default govim highlight rules to, for example,
change the color of error signs. Update your overrides according
to the new highlight names.

Override the following highlights in your .vimrc to alter the
appearance of text properties and signs.
- GOVIMSignErr, GOVIMSignWarn, GOVIMSignInfo & GOVIMSignHint
- GOVIMErr, GOVIMWarn, GOVIMInfo & GOVIMHint