tests: create gists from artefacts (logs) of failed CI tests
Very often we get "useful" failures on CI that are hard to reproduce
locally. Given the headless nature of CI it's impossible to then extract
the log files from those failures. This is very frustrating.

This commit causes test failures on Travis to create a gist with a
base64 encoded .tar.gz of the log files from all tests that have, to
that point, run. For convenience, a command is echo-ed that can be
copy-pasted into your shell that will change to a new temp dir and
untar the log files ready for inspection.