cmd/govim: create a test demonstrating failing completion (#659)
As previously discovered (and also raised in #658), 11eb9c7 caused
completion to break in govim because placeholder values are now included
in the completion result. This has now been fixed in:

  https://go-review.googlesource.com/c/tools/+/213640

However, because of a bug introduced in:

  https://go-review.googlesource.com/c/tools/+/212102

we can't simply upgrade to the latest x/tools and gopls.

In any case we want to add a test to demonstrate the problem and prevent
and future regressions. A subsequent PR will upgrade the version of
x/tools and gopls to a commit (in a temporary fork) that has the fix for
the completion bug but does not have the bug for the parallel initial
load.

Updates #658