cmd/govim: fix GOVIMQuickfixDiagnostics command (#662)
Per #657, after 2975a8f `:GOVIMQuickfixDiagnostics` is broken as long as
QuickfixAutoDiagnostics is disabled. This was an unintentional break. We
managed to break this command because the command was not actually
tested.

Fix both of these points by fixing the bug itself and also adding a test
to prevent further breakage.

Fixes #657