cmd/govim: Do not increase version if buffer content didn't change
We currently blindly increase buffer version when vim triggers
BufReadPost or BufNewFile (and a didOpen or didChange request to
gopls). Unloading and loading a buffer that is open also triggers
a BufReadPost, and if the content didn't change we shouldn't
increase version.

This change also skip redefining highlights if the version is
indeed increased since the last known diagnostics could have
changed.

Updates #655