testdriver: add vimexprwait to simplify wait assertions (#678)
vimexprwait repeatedly evaluates a Vim expression (where the expression
is like that passed to the vim expr command) until the result matches
the contents of a golden file.

    vimexprwait golden.file expr

e.g.

    vimexprwait errors.golden getqflist()

will wait until the quickfix list contents match the contents of
errors.golden.

By default vimexprwait JSON indent-formats the result
(this can be turned off with -noindent).

By default vimexprwait waits (with exponential backoff) for up to
testdriver.DefaultErrLogMatchWait.

As part of this change we also simplify all tests that previously used
an approach that was a mix of errlogmatch, vim expr, copen etc. The
semantics of the changed tests have not changed, only the way in which
we assert.

For file watcher-based tests, we move away from waiting for diagnostics
for a file spotted by the watcher to instead specifically looking for
didOpen/didChange events being sent to gopls. Given the happens-before
relationship that didOpen/didChange affords us, again this doesn't
result in any change in semantics.

Given vimexprwait can depend on the value of GOVIM_ERRLOGMATCH_WAIT, so
we also rename DEFAULT_ERRLOGMATCH_WAIT to GOVIM_ERRLOGMATCH_WAIT and
adjust all comments/sleeps accordingly.

As part of this we make a copy of
github.com/rogpeppe/go-internal/internal/textutil for our local needs.

Also remove the current "hack" whereby we copen/cclose prior to running
any script in order to determine the quickfix buffer number; this hack
is no longer required because we no longer copen in order to dump the
quickfix contents - instead we use vimexprwait. This means that some
previous golden values need to change as the first open file will be in
bufnr 1, and window numbers will start from 1000, with the next buffer
being bufn2 and the next window being 1001.