deps: use temporary fork of x/tools and gopls
In d80f0e21 we upgraded to the latest x/tools and gopls. But as is
documented in golang.org/issue/36601 there is a bug whereby reverting a
file to an earlier state results in no diagnostics where they are in
fact expected.

Fix this by using a temporary fork of x/tools and gopls where CL 214586
is partially reverted.

Also add tests to ensure the various permutations of transitions to/from
error state and back work as expected.